requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.2
opencv-python>=4.6.0
pillow>=8.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
pyyaml>=6.0
wandb>=0.15.0
codecarbon>=2.1.0
tqdm>=4.64.0
scipy>=1.7.0
tensorboard>=2.8.0
requests>=2.23.0
scipy>=1.4.1
psutil
py-cpuinfo
thop>=0.1.1